+++ packages/expo-location/src/LocationEventEmitter.ts
@@ -1,5 +1,5 @@
-import { EventEmitter } from 'expo-modules-core';
+import { LegacyEventEmitter } from 'expo-modules-core';
 
 import ExpoLocation from './ExpoLocation';
 
-export const LocationEventEmitter = new EventEmitter(ExpoLocation);
+export const LocationEventEmitter = new LegacyEventEmitter(ExpoLocation);