+++ packages/expo-notifications/src/NotificationsEmitter.ts
@@ -1,11 +1,11 @@
-import { EventEmitter, Subscription, UnavailabilityError } from 'expo-modules-core';
+import { LegacyEventEmitter, type EventSubscription, UnavailabilityError } from 'expo-modules-core';
 
 import { Notification, NotificationResponse } from './Notifications.types';
 import NotificationsEmitterModule from './NotificationsEmitterModule';
 import { mapNotificationResponse } from './utils/mapNotificationResponse';
 
 // Web uses SyntheticEventEmitter
-const emitter = new EventEmitter(NotificationsEmitterModule);
+const emitter = new LegacyEventEmitter(NotificationsEmitterModule);
 
 const didReceiveNotificationEventName = 'onDidReceiveNotification';
 const didDropNotificationsEventName = 'onNotificationsDeleted';
@@ -40,7 +40,7 @@ export const DEFAULT_ACTION_IDENTIFIER = 'expo.modules.notifications.actions.DEF
  */
 export function addNotificationReceivedListener(
   listener: (event: Notification) => void
-): Subscription {
+): EventSubscription {
   return emitter.addListener<Notification>(didReceiveNotificationEventName, listener);
 }
 
@@ -52,7 +52,7 @@ export function addNotificationReceivedListener(
  * @return A [`Subscription`](#subscription) object represents the subscription of the provided listener.
  * @header listen
  */
-export function addNotificationsDroppedListener(listener: () => void): Subscription {
+export function addNotificationsDroppedListener(listener: () => void): EventSubscription {
   return emitter.addListener<void>(didDropNotificationsEventName, listener);
 }
 
@@ -84,7 +84,7 @@ export function addNotificationsDroppedListener(listener: () => void): Subscript
  */
 export function addNotificationResponseReceivedListener(
   listener: (event: NotificationResponse) => void
-): Subscription {
+): EventSubscription {
   return emitter.addListener<NotificationResponse>(
     didReceiveNotificationResponseEventName,
     (response: NotificationResponse) => {
@@ -99,8 +99,8 @@ export function addNotificationResponseReceivedListener(
  * @param subscription A subscription returned by `addNotificationListener` method.
  * @header listen
  */
-export function removeNotificationSubscription(subscription: Subscription) {
-  emitter.removeSubscription(subscription);
+export function removeNotificationSubscription(subscription: EventSubscription) {
+  subscription.remove();
 }
 
 // @docsMissing