+++ packages/expo-sqlite/src/SQLiteDatabase.ts
@@ -1,4 +1,4 @@
-import { EventEmitter, Subscription } from 'expo-modules-core';
+import { type EventSubscription } from 'expo-modules-core';
 
 import ExpoSQLite from './ExpoSQLiteNext';
 import { NativeDatabase, SQLiteOpenOptions } from './NativeDatabase';
@@ -12,8 +12,6 @@ import {
 } from './SQLiteStatement';
 
 export { SQLiteOpenOptions };
-
-const emitter = new EventEmitter(ExpoSQLite);
 
 /**
  * A SQLite database.
@@ -523,8 +521,8 @@ export type DatabaseChangeEvent = {
  */
 export function addDatabaseChangeListener(
   listener: (event: DatabaseChangeEvent) => void
-): Subscription {
-  return emitter.addListener('onDatabaseChange', listener);
+): EventSubscription {
+  return ExpoSQLite.addListener('onDatabaseChange', listener);
 }
 
 /**