+++ packages/expo-updates/src/UpdatesEmitter.ts
@@ -1,12 +1,10 @@
-import { EventEmitter as NativeEventEmitter } from 'expo-modules-core';
 import { EventEmitter as JsEventEmitter } from 'fbemitter';
 
 import ExpoUpdatesModule from './ExpoUpdates';
 import { transformNativeStateMachineContext } from './Updates';
 import type { UpdatesNativeStateChangeEvent } from './Updates.types';
 
-const _nativeEventEmitter = new NativeEventEmitter(ExpoUpdatesModule);
-_nativeEventEmitter.addListener('Expo.nativeUpdatesStateChangeEvent', _emitNativeStateChangeEvent);
+ExpoUpdatesModule.addListener('Expo.nativeUpdatesStateChangeEvent', _emitNativeStateChangeEvent);
 
 let _jsEventEmitter: JsEventEmitter | null = null;
 function _getJsEventEmitter(): JsEventEmitter {